pyproject.toml
[tool.poetry]
name = "madlad-finetuning"
version = "0.1.0"
description = ""
authors = ["Mark <woodwardmw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
transformers = "^4.44.2"
accelerate = "^0.34.2"
sentencepiece = "^0.2.0"
protobuf = "^5.28.2"
torch = "^2.4.1"
clearml = "^1.16.4"
datasets = ">=2.0.0,<3.0.0"
networkx = "<3.2"
evaluate = "^0.4.3"
sacrebleu = "^2.4.3"
peft = "^0.12.0"
python-dotenv = "^1.0.1"
sil-machine = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
